--- entities/movie/api/queries/get-movie.ts
@@ -1,10 +1,0 @@
-import { mapTMDBMovieDetail } from "../mapper";
-import { tmdbFetch } from "../tmdb-client";
-import { TMDBMovieDetail } from "../tmdb-types";
-
-export async function getMovieDetails(id: string) {
-  const data = await tmdbFetch<TMDBMovieDetail>(`/movie/${id}`, {
-    language: "ko-KR",
-  });
-  return mapTMDBMovieDetail(data);
-}
