+++ entities/movie/index.ts
@@ -1,9 +1,7 @@
 export type { Movie, MovieDetail, HeroMovie, CastMember, Director, MovieCredits } from "./model/types";
 
-export { getPopularMovies } from "./api/queries/get-popular";
-export { getMovieDetails } from "./api/queries/get-movie";
-export { getMovieTrailer } from "./api/queries/get-movie-trailer";
-export { getMovieCredits } from "./api/queries/get-credits";
-export { getSimilarMovies } from "./api/queries/get-similar-movie";
-
-export { buildImageUrl } from "./lib/image-url";
+export { getPopularMovies } from "./api/getPopularMovies";
+export { getMovie } from "./api/getMovie";
+export { getMovieTrailer } from "./api/getMovieTrailer";
+export { getMovieCredits } from "./api/getMovieCredits";
+export { getSimilarMovies } from "./api/getSimilarMovies";
